Integrate rocketmq-related content into storage module
