+++ adapter/runtimer/pom.xml
@@ -43,11 +43,6 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>org.apache.rocketmq</groupId>
-            <artifactId>rocketmq-acl</artifactId>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>io.openmessaging</groupId>
             <artifactId>openmessaging-connector</artifactId>
             <version>0.1.2</version>
