+++ adapter/storage/src/main/java/org/apache/rocketmq/eventbridge/adapter/storage/rocketmq/runtimer/consumer/RetryPolicy.java
@@ -15,32 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.rocketmq.eventbridge.adapter.runtimer.boot.listener.rocketmq;
-
-import org.apache.rocketmq.client.exception.MQClientException;
-import org.apache.rocketmq.common.message.MessageExt;
-
-import java.time.Duration;
-import java.util.List;
+package org.apache.rocketmq.eventbridge.adapter.storage.rocketmq.runtimer.consumer;
 
 /**
  * @Author changfeng
- * @Date 2023/4/9 10:09 上午
+ * @Date 2023/4/9 10:10 上午
  */
-public interface LitePullConsumer {
-    void startup() throws MQClientException;
-
-    void shutdown();
-
-    void attachTopic(String topic, String tag);
-
-    List<MessageExt> poll(int pullBatchSize, Duration timeout);
-
-    void commit(List<String> messageIdList);
-
-    void setSockProxyJson(String proxyJson);
-
-    void subscribe(String topic);
-
-    void unsubscribe(String topic);
+public interface RetryPolicy {
+    long nextDelayDuration();
 }
